Updating polybar scripts
Mainly for Desktop PC, Laptop is not using this stuff

The point is, I've created a script called UPDATE-SCRIPTS.sh
and a static_scripts.patch in order to clone the raw updated
scripts and apply my personal customizations with icons or
fixes.

Read the comments on UPDATE-SCRIPTS.sh for more info.